backend/requirements.txt
asgiref==3.8.1
black==25.1.0
Django==4.2.20
djangorestframework==3.16.0
django-filter==25.1
djoser==2.3.1
psycopg2-binary==2.9.9
isort==6.0.1
flake8==7.2.0
autopep8==2.3.2
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
Pillow==11.3.0